requirements.txt
camelot-py==0.7.2
peewee==3.9.5
pandas==0.24.2
psycopg2==2.8.2
requests==2.22.0